create test_transformations.py and update it
