+++ tests/unit_tests/utils/test_general.py
@@ -9,13 +9,10 @@ from copy import copy
 from pathlib import Path
 
 import astropy.units as u
-import numpy as np
 import pytest
-from astropy.coordinates.errors import UnitsError
 from astropy.io.misc import yaml
 
 import simtools.utils.general as gen
-import simtools.utils.transformations as transf
 from simtools.utils.general import (
     InvalidConfigData,
     InvalidConfigEntry,
@@ -306,87 +303,6 @@ def test_change_dict_keys_case(caplog) -> None:
         assert "Input is not a proper dictionary" in caplog.text
 
 
-def test_rotate_telescope_position(caplog) -> None:
-    x = np.array([-10, -10, 10, 10]).astype(float)
-    y = np.array([-10.0, 10.0, -10.0, 10.0]).astype(float)
-    angle_deg = 30 * u.deg
-    x_rot_manual = np.array([-3.7, -13.7, 13.7, 3.7])
-    y_rot_manual = np.array([-13.7, 3.7, -3.7, 13.7])
-
-    def check_results(x_to_test, y_to_test, x_right, y_right, angle, theta=0 * u.deg):
-        x_rot, y_rot = transf.rotate(x_to_test, y_to_test, angle, theta)
-        x_rot, y_rot = np.around(x_rot, 1), np.around(y_rot, 1)
-        if not isinstance(x_right, (list, np.ndarray)):
-            x_right = [x_right]
-        if not isinstance(y_right, (list, np.ndarray)):
-            y_right = [y_right]
-        for element, _ in enumerate(x_right):
-            assert x_right[element] == x_rot[element]
-            assert y_right[element] == y_rot[element]
-
-    # Testing without units
-    check_results(x, y, x_rot_manual, y_rot_manual, angle_deg)
-
-    # Testing with scalars
-    check_results(-10.0, -10.0, -3.7, -13.7, 30 * u.deg)
-
-    x_new_array, y_new_array = x * u.m, y * u.m
-    x_rot_new_array, y_rot_new_array = x_rot_manual * u.m, y_rot_manual * u.m
-
-    # Testing with units
-    check_results(x_new_array, y_new_array, x_rot_new_array, y_rot_new_array, angle_deg)
-
-    # Testing with radians
-    check_results(x_new_array, y_new_array, x_rot_new_array, y_rot_new_array, angle_deg.to(u.rad))
-
-    # Testing rotation in theta, around Y (3D)
-    x_rot_theta_manual = np.array([-2.6, -9.7, 9.7, 2.6])
-    y_rot_theta_manual = np.array([-13.7, 3.7, -3.7, 13.7])
-    check_results(x, y, x_rot_theta_manual, y_rot_theta_manual, angle_deg, 45 * u.deg)
-
-    with pytest.raises(TypeError):
-        transf.rotate(x, y[0], angle_deg)
-    with pytest.raises(TypeError):
-        transf.rotate("1", "2", angle_deg)
-        assert "x and y types are not valid! Cannot perform transformation" in caplog.text
-    with pytest.raises(TypeError):
-        transf.rotate(str(x[0]), y[0], angle_deg)
-    with pytest.raises(TypeError):
-        transf.rotate(u.Quantity(10), 10, angle_deg)
-    with pytest.raises(TypeError):
-        transf.rotate(x[0], str(y[0]), angle_deg)
-    with pytest.raises(RuntimeError):
-        transf.rotate(x[:-1], y, angle_deg)
-    with pytest.raises(UnitsError):
-        transf.rotate(x_new_array.to(u.cm), y_new_array, angle_deg)
-    with pytest.raises(u.core.UnitsError):
-        transf.rotate(x_new_array, y_new_array, 30 * u.m)
-
-
-def test_convert_2D_to_radial_distr(caplog) -> None:
-    # Test normal functioning
-    max_dist = 100
-    bins = 100
-    step = max_dist / bins
-    xaxis = np.arange(-max_dist, max_dist, step)
-    yaxis = np.arange(-max_dist, max_dist, step)
-    x2d, y2d = np.meshgrid(xaxis, yaxis)
-    distance_to_center_2D = np.sqrt((x2d) ** 2 + (y2d) ** 2)
-
-    distance_to_center_1D, radial_bin_edges = transf.convert_2D_to_radial_distr(
-        distance_to_center_2D, xaxis, yaxis, bins=bins, max_dist=max_dist
-    )
-    difference = radial_bin_edges[:-1] - distance_to_center_1D
-    assert pytest.approx(difference[:-1], abs=1) == 0  # last value deviates
-
-    # Test warning in caplog
-    transf.convert_2D_to_radial_distr(
-        distance_to_center_2D, xaxis, yaxis, bins=4 * bins, max_dist=max_dist
-    )
-    msg = "The histogram with number of bins"
-    assert msg in caplog.text
-
-
 def test_save_dict_to_file(tmp_test_directory, caplog) -> None:
     paths = ["test_file", "test_file.yml"]
     example_dict = {"key": 12}
@@ -517,42 +433,6 @@ def test_set_default_kwargs() -> None:
     in_kwargs = {"a": 1, "b": 2}
     out_kwargs = gen.set_default_kwargs(in_kwargs, c=3, d=4)
     assert out_kwargs == {"a": 1, "b": 2, "c": 3, "d": 4}
-
-
-def test_sort_arrays() -> None:
-    """
-    Test the sort_arrays function.
-    """
-
-    # Test with no arguments.
-    args = []
-    new_args = transf.sort_arrays(*args)
-    assert not new_args
-
-    # Test with one argument.
-    args = [list(range(10))]
-    new_args = transf.sort_arrays(*args)
-    assert new_args == [list(range(10))]
-
-    # Test with multiple arguments.
-    args = [list(range(10)), list(range(10, 20))]
-    new_args = transf.sort_arrays(*args)
-    assert new_args == [list(range(10)), list(range(10, 20))]
-
-    # Test with arguments of different lengths.
-    args = [list(range(10)), list(range(5))]
-    new_args = transf.sort_arrays(*args)
-    assert new_args == [list(range(10)), list(range(5))]
-
-    # Test with arguments that are not arrays.
-    args = [1, 2, 3]
-    with pytest.raises(TypeError):
-        transf.sort_arrays(*args)
-
-    # Test with the input array not in the right order.
-    args = [list(reversed(range(10)))]
-    new_args = transf.sort_arrays(*args)
-    assert new_args == [list(range(10))]
 
 
 def test_collect_final_lines(tmp_test_directory) -> None:
